Main/VBoxSVC: enable -Wconversion plus a couple of fixes (all harmless)
svn:sync-xref-src-repo-rev: r108908

+++ src/VBox/Main/src-all/DisplayUtils.cpp
@@ -1,4 +1,4 @@
-/* $Id: DisplayUtils.cpp 54792 2015-03-16 16:04:24Z vadim.galitsyn@oracle.com $ */
+/* $Id: DisplayUtils.cpp 62363 2016-07-20 15:45:58Z noreply@oracle.com $ */
 /** @file
  * Implementation of IDisplay helpers.
  */
@@ -69,7 +69,7 @@ int readSavedDisplayScreenshot(const Utf8Str &strStateFilePath, uint32_t u32Type
                     {
                         if (cbBlock > 2 * sizeof(uint32_t))
                         {
-                            cbData = cbBlock - 2 * sizeof(uint32_t);
+                            cbData = (uint32_t)(cbBlock - 2 * sizeof(uint32_t));
                             pu8Data = (uint8_t *)RTMemAlloc(cbData);
                             if (pu8Data == NULL)
                             {
